# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStats {
  totalUsers: number;
  totalWarnings: number;
  activePolls: number;
  totalGuilds: number;
  topLevel: number;
  totalEconomyCoins: number;
}

export interface LeaderboardEntry {
  userId: string;
  xp: number;
  level: number;
  /** @nullable */
  messageCount?: number | null;
  /** @nullable */
  voiceMinutes?: number | null;
  rank: number;
}

export interface EconomyEntry {
  userId: string;
  balance: number;
  rank: number;
}

export interface Warning {
  id: string;
  guildId: string;
  userId: string;
  moderatorId: string;
  reason: string;
  timestamp: number;
}

export type PollVotes = { [key: string]: number };

export interface Poll {
  id: string;
  guildId: string;
  question: string;
  options: string[];
  votes: PollVotes;
  ended: boolean;
  createdAt: number;
  creatorId: string;
}

export interface Guild {
  id: string;
  userCount: number;
}

export interface WarningInput {
  guildId: string;
  userId: string;
  moderatorId: string;
  reason: string;
}

export interface UserProfile {
  id: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface CasinoDailyInput {
  guildId: string;
  userId: string;
}

export interface CasinoBetInput {
  guildId: string;
  userId: string;
  bet: number;
}

export interface CasinoRouletteInput {
  guildId: string;
  userId: string;
  bet: number;
  choice: string;
}

export type CasinoActionInputAction =
  (typeof CasinoActionInputAction)[keyof typeof CasinoActionInputAction];

export const CasinoActionInputAction = {
  hit: "hit",
  stand: "stand",
  double: "double",
} as const;

export interface CasinoActionInput {
  userId: string;
  action: CasinoActionInputAction;
}

export interface CasinoConfig {
  currency: string;
  minBet: number;
  maxBet: number;
  startingBalance: number;
  dailyAmount: number;
  slotsJackpotMultiplier: number;
}

export interface CasinoBalance {
  balance: number;
  streak: number;
  canDaily: boolean;
  nextDailyMs: number;
}

export interface DailyClaimResult {
  balance: number;
  earned: number;
  streak: number;
  bonusAmount: number;
}

export interface SlotsResult {
  reels: string[];
  multiplier: number;
  winnings: number;
  net: number;
  balance: number;
  resultType: string;
}

export interface RouletteResult {
  roll: number;
  color: string;
  won: boolean;
  multiplier: number;
  label: string;
  winnings: number;
  net: number;
  balance: number;
}

export interface BlackjackCard {
  rank: string;
  suit: string;
}

export interface BlackjackState {
  status: string;
  playerHand: BlackjackCard[];
  dealerHand: BlackjackCard[];
  playerValue: number;
  dealerValue: number;
  bet: number;
  payout: number;
  balance: number;
  canDouble: boolean;
}

export type GetLeaderboardParams = {
  guildId: string;
  limit?: number;
};

export type GetEconomyParams = {
  guildId: string;
  limit?: number;
};

export type GetWarningsParams = {
  guildId: string;
};

export type ResolveUsersParams = {
  /**
   * Comma-separated list of Discord user IDs
   */
  ids: string;
};

export type GetCasinoConfigParams = {
  guildId: string;
};

export type GetCasinoBalanceParams = {
  guildId: string;
  userId: string;
};

export type GetPollsParams = {
  guildId: string;
};