# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bot
    description: Discord bot data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bot/stats:
    get:
      operationId: getBotStats
      tags: [bot]
      summary: Overall bot statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStats"

  /bot/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [bot]
      summary: XP leaderboard for a guild
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  /bot/economy:
    get:
      operationId: getEconomy
      tags: [bot]
      summary: Economy leaderboard for a guild
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Economy entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EconomyEntry"

  /bot/warnings:
    get:
      operationId: getWarnings
      tags: [bot]
      summary: Warnings list for a guild
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warnings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warning"
    post:
      operationId: createWarning
      tags: [bot]
      summary: Add a warning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarningInput"
      responses:
        "201":
          description: Warning created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warning"

  /bot/warnings/{warningId}:
    delete:
      operationId: deleteWarning
      tags: [bot]
      summary: Delete a warning by id
      parameters:
        - name: warningId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bot/resolve:
    get:
      operationId: resolveUsers
      tags: [bot]
      summary: Resolve Discord user IDs to profile info
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of Discord user IDs
      responses:
        "200":
          description: User profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  /bot/casino/config:
    get:
      operationId: getCasinoConfig
      tags: [bot]
      summary: Casino configuration for a guild
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Casino config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoConfig"

  /bot/casino/balance:
    get:
      operationId: getCasinoBalance
      tags: [bot]
      summary: User balance and daily status
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasinoBalance"

  /bot/casino/daily:
    post:
      operationId: claimCasinoDaily
      tags: [bot]
      summary: Claim daily reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoDailyInput"
      responses:
        "200":
          description: Daily claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyClaimResult"

  /bot/casino/slots:
    post:
      operationId: playCasinoSlots
      tags: [bot]
      summary: Play slots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoBetInput"
      responses:
        "200":
          description: Slots result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsResult"

  /bot/casino/roulette:
    post:
      operationId: playCasinoRoulette
      tags: [bot]
      summary: Play roulette
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoRouletteInput"
      responses:
        "200":
          description: Roulette result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouletteResult"

  /bot/casino/blackjack/start:
    post:
      operationId: startCasinoBlackjack
      tags: [bot]
      summary: Start a blackjack game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoBetInput"
      responses:
        "200":
          description: Blackjack state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackjackState"

  /bot/casino/blackjack/action:
    post:
      operationId: actionCasinoBlackjack
      tags: [bot]
      summary: Blackjack action (hit, stand, double)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasinoActionInput"
      responses:
        "200":
          description: Updated blackjack state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackjackState"

  /bot/polls:
    get:
      operationId: getPolls
      tags: [bot]
      summary: Polls for a guild
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"

  /bot/guilds:
    get:
      operationId: getGuilds
      tags: [bot]
      summary: List of guilds tracked by the bot
      responses:
        "200":
          description: Guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BotStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalWarnings:
          type: integer
        activePolls:
          type: integer
        totalGuilds:
          type: integer
        topLevel:
          type: integer
        totalEconomyCoins:
          type: integer
      required:
        - totalUsers
        - totalWarnings
        - activePolls
        - totalGuilds
        - topLevel
        - totalEconomyCoins

    LeaderboardEntry:
      type: object
      properties:
        userId:
          type: string
        xp:
          type: integer
        level:
          type: integer
        messageCount:
          type: ["integer", "null"]
        voiceMinutes:
          type: ["integer", "null"]
        rank:
          type: integer
      required:
        - userId
        - xp
        - level
        - rank

    EconomyEntry:
      type: object
      properties:
        userId:
          type: string
        balance:
          type: integer
        rank:
          type: integer
      required:
        - userId
        - balance
        - rank

    Warning:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        userId:
          type: string
        moderatorId:
          type: string
        reason:
          type: string
        timestamp:
          type: integer
      required:
        - id
        - guildId
        - userId
        - moderatorId
        - reason
        - timestamp

    Poll:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        votes:
          type: object
          additionalProperties:
            type: integer
        ended:
          type: boolean
        createdAt:
          type: integer
        creatorId:
          type: string
      required:
        - id
        - guildId
        - question
        - options
        - votes
        - ended
        - createdAt
        - creatorId

    Guild:
      type: object
      properties:
        id:
          type: string
        userCount:
          type: integer
      required:
        - id
        - userCount

    WarningInput:
      type: object
      properties:
        guildId:
          type: string
        userId:
          type: string
        moderatorId:
          type: string
        reason:
          type: string
      required:
        - guildId
        - userId
        - moderatorId
        - reason

    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username

    CasinoDailyInput:
      type: object
      properties:
        guildId:
          type: string
        userId:
          type: string
      required: [guildId, userId]

    CasinoBetInput:
      type: object
      properties:
        guildId:
          type: string
        userId:
          type: string
        bet:
          type: integer
      required: [guildId, userId, bet]

    CasinoRouletteInput:
      type: object
      properties:
        guildId:
          type: string
        userId:
          type: string
        bet:
          type: integer
        choice:
          type: string
      required: [guildId, userId, bet, choice]

    CasinoActionInput:
      type: object
      properties:
        userId:
          type: string
        action:
          type: string
          enum: [hit, stand, double]
      required: [userId, action]

    CasinoConfig:
      type: object
      properties:
        currency:
          type: string
        minBet:
          type: integer
        maxBet:
          type: integer
        startingBalance:
          type: integer
        dailyAmount:
          type: integer
        slotsJackpotMultiplier:
          type: integer
      required:
        - currency
        - minBet
        - maxBet
        - startingBalance
        - dailyAmount
        - slotsJackpotMultiplier

    CasinoBalance:
      type: object
      properties:
        balance:
          type: integer
        streak:
          type: integer
        canDaily:
          type: boolean
        nextDailyMs:
          type: integer
      required:
        - balance
        - streak
        - canDaily
        - nextDailyMs

    DailyClaimResult:
      type: object
      properties:
        balance:
          type: integer
        earned:
          type: integer
        streak:
          type: integer
        bonusAmount:
          type: integer
      required:
        - balance
        - earned
        - streak
        - bonusAmount

    SlotsResult:
      type: object
      properties:
        reels:
          type: array
          items:
            type: string
        multiplier:
          type: number
        winnings:
          type: integer
        net:
          type: integer
        balance:
          type: integer
        resultType:
          type: string
      required:
        - reels
        - multiplier
        - winnings
        - net
        - balance
        - resultType

    RouletteResult:
      type: object
      properties:
        roll:
          type: integer
        color:
          type: string
        won:
          type: boolean
        multiplier:
          type: integer
        label:
          type: string
        winnings:
          type: integer
        net:
          type: integer
        balance:
          type: integer
      required:
        - roll
        - color
        - won
        - multiplier
        - label
        - winnings
        - net
        - balance

    BlackjackCard:
      type: object
      properties:
        rank:
          type: string
        suit:
          type: string
      required:
        - rank
        - suit

    BlackjackState:
      type: object
      properties:
        status:
          type: string
        playerHand:
          type: array
          items:
            $ref: "#/components/schemas/BlackjackCard"
        dealerHand:
          type: array
          items:
            $ref: "#/components/schemas/BlackjackCard"
        playerValue:
          type: integer
        dealerValue:
          type: integer
        bet:
          type: integer
        payout:
          type: integer
        balance:
          type: integer
        canDouble:
          type: boolean
      required:
        - status
        - playerHand
        - dealerHand
        - playerValue
        - dealerValue
        - bet
        - payout
        - balance
        - canDouble