# lib/api-zod/src/generated/types/blackjackState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlackjackCard } from "./blackjackCard";

export interface BlackjackState {
  status: string;
  playerHand: BlackjackCard[];
  dealerHand: BlackjackCard[];
  playerValue: number;
  dealerValue: number;
  bet: number;
  payout: number;
  balance: number;
  canDouble: boolean;
}