# lib/api-zod/src/generated/types/casinoActionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CasinoActionInputAction } from "./casinoActionInputAction";

export interface CasinoActionInput {
  userId: string;
  action: CasinoActionInputAction;
}