# lib/api-zod/src/generated/types/casinoConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CasinoConfig {
  currency: string;
  minBet: number;
  maxBet: number;
  startingBalance: number;
  dailyAmount: number;
  slotsJackpotMultiplier: number;
}